+++ src/tnn_numeric.h
@@ -3,7 +3,7 @@
  * Version 0.1, 02/25/2012
  *
  * This header defines the following functions:
- * tnn_error tnn_numeric_v2m(gsl_vector *v, gsl_matrix *m, int size1, int size2);
+ * tnn_error tnn_numeric_v2m(gsl_vector *v, gsl_matrix *m, size_t size1, size_t size2);
  */
 
 #include <gsl/gsl_vector.h>
@@ -14,6 +14,6 @@
 #define TNN_NUMERIC_H
 
 //Vector to matrix conversion
-tnn_error tnn_numeric_v2m(gsl_vector *v, gsl_matrix *m, int size1, int size2);
+tnn_error tnn_numeric_v2m(gsl_vector *v, gsl_matrix *m, size_t size1, size_t size2);
 
 #endif //TNN_NUMERIC_H
