+++ src/tnn_state.c
@@ -3,24 +3,26 @@
  * Version 0.1, 02/19/2012
  *
  * This header implements the following functions:
- * int tnn_state_init(tnn_state *s);
+ * tnn_error tnn_state_init(tnn_state *s, size_t n);
+ * tnn_error tnn_state_debug(tnn_state *s);
  */
 
+#include <stddef.h>
 #include <stdbool.h>
 #include <stdio.h>
 #include <gsl/gsl_vector.h>
 #include <tnn_state.h>
 #include <tnn_error.h>
 
-tnn_error tnn_state_init(tnn_state *s, int n){
+tnn_error tnn_state_init(tnn_state *s, size_t n){
   s->valid = false;
   s->size = n;
   return TNN_ERROR_SUCCESS;
 }
 
 tnn_error tnn_state_debug(tnn_state *s){
-  int i;
-  printf("state = %p, size = %d, valid = %c, prev = %p, next = %p\n", s, s->size, s->valid == true?'T':'F', s->prev, s->next);
+  size_t i;
+  printf("state = %p, size = %ld, valid = %c, prev = %p, next = %p\n", s, s->size, s->valid == true?'T':'F', s->prev, s->next);
   if(s->valid == true){
     printf("x:");
     for(i = 0; i < s->size; i = i + 1){
